08/54: Sync 'Kurzerfassungs' fields where standard fields are missing
There are the traditional fields methods, population and results. In
addition there are additional fields that capture dedicated parts of
those fields ('Kurzerfassungs' fields). When synching the core records
into public, the standard fields have precendence if they are populated.
If however one of those fields hold a null or blank value, SciPaMaTo
fills in the short fields into the respective target fields.

The mapping is as follows:

 * Methods: methodStudyDesign - methodOutcome - methodStatistics -
   methodconfounders

 * Population: populationPlace - populationParticipants -
   populationDuration

 * Result: resultExposureRange - resultEffectEstimate -
   resultMeasuredOutcome

It is expected that the users will request some amendments in this
procedure. This is only a first shot to demonstrate what's possible.