Fix deletion of search order
After having deleted a search order, the select box loaded the next one but the panel did not reload its model properly. Hitting
delete on the next search order tried to delete the one that has already been deleted previously.

After deleting a search order, the SearchOrderPage is reloaded from scratch, hence the panels model is blank as well as the
search order combobox. In order to delete another search order it has to be selected explicitly - keeping the combobox and the panel
model in sync