Use improved TestAjaxRequestHandler to considerably speed up the tests
Use LinkedHashSet instead of LinkedList to store the instances of
IListener. This improves the time complexity in method addListener
for the contains method from O(n) to O(1). This brings down the time
spent for running the `gradlew check` command for the entire project
from several hours to 20 minutes again (with mockk - mockito was not
affected that much).