fix: [#260] Log in to dockerhub in github actions (#261)
With dockerhub rate-limiting in effect, pulling docker images anonymously faces the risk of failing
with a 404. This PR adds two explicit steps to a) log in to dockerhub and b) download the postgresql
image before running gradle. jooq-modelator and test-containers will not need to pull the image
anonymously as it is already present in the Linux container.
Downside: The postgresql image is defined in two places.